+++ src/BizHawk.Emulation.Cores/Consoles/Nintendo/GBHawk/Mappers/Mapper_MBC1.cs
@@ -58,12 +58,16 @@ namespace BizHawk.Emulation.Cores.Nintendo.GBHawk
 				}
 				else
 				{
-					return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8) ? 0xFF : Core.bus_value;
+					return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8)
+						? (byte) 0xFF
+						: Core.bus_value;
 				}
 			}
 			else
 			{
-				return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8) ? 0xFF : Core.bus_value;
+				return Core.cpu.TotalExecutedCycles > (Core.bus_access_time + 8)
+					? (byte) 0xFF
+					: Core.bus_value;
 			}
 		}
